+++ usb_descriptor.h
@@ -60,7 +60,7 @@ static const struct usb_device_descriptor device_descriptor = {
     .bMaxPacketSize0 = 64,    // Max packet size for ep0
     .idVendor = VENDOR_ID,    // Your vendor id
     .idProduct = PRODUCT_ID,  // Your product ID
-    .bcdDevice = PICODAC_EAC3_PASSTHROUGH ? 0x0110 : 0x0108,
+    .bcdDevice = PICODAC_EAC3_PASSTHROUGH ? 0x0111 : 0x0108,
     .iManufacturer = 0,       // Manufacturer string index
     .iProduct = 0,            // Product string index
     .iSerialNumber = 0,       // No serial number
@@ -122,11 +122,6 @@ struct configuration_descriptor {
         cs_ac_output_terminal;
     struct usb_class_specific_ac_feature_unit_descriptor_stereo
         cs_ac_feature_unit;
-#if PICODAC_EAC3_PASSTHROUGH
-    struct usb_class_specific_ac_clock_source_descriptor eac3_clock;
-    struct usb_class_specific_ac_input_terminal_descriptor eac3_input;
-    struct usb_class_specific_ac_output_terminal_descriptor eac3_output;
-#endif
   } __attribute__((packed)) ac;
   struct as {
     struct as_alt0 {
@@ -279,27 +274,6 @@ struct configuration_descriptor {
                     .iFeature = 0,
                 },
         },
-#if PICODAC_EAC3_PASSTHROUGH
-    .ac.eac3_clock = {
-        .bLength = sizeof(struct usb_class_specific_ac_clock_source_descriptor),
-        .bDescriptorType = USB_DT_CS_INTERFACE, .bDescriptorSubtype = 0x0A,
-        .bClockID = AUDIO_CONTROL_ID_EAC3_CLOCK,
-        .bmAttributes = 1, .bmControls = 0x05, // fixed internal clock, freq/validity read-only
-    },
-    .ac.eac3_input = {
-        .bLength = sizeof(struct usb_class_specific_ac_input_terminal_descriptor),
-        .bDescriptorType = USB_DT_CS_INTERFACE, .bDescriptorSubtype = 2,
-        .bTerminalID = AUDIO_CONTROL_ID_EAC3_INPUT, .wTerminalType = 0x0101,
-        .bCSourceID = AUDIO_CONTROL_ID_EAC3_CLOCK, .bNrChannels = 2, .bmChannelConfig = 3,
-    },
-    .ac.eac3_output = {
-        .bLength = sizeof(struct usb_class_specific_ac_output_terminal_descriptor),
-        .bDescriptorType = USB_DT_CS_INTERFACE, .bDescriptorSubtype = 3,
-        .bTerminalID = AUDIO_CONTROL_ID_EAC3_OUTPUT, .wTerminalType = 0x0605,
-        .bSourceID = AUDIO_CONTROL_ID_EAC3_INPUT,
-        .bCSourceID = AUDIO_CONTROL_ID_EAC3_CLOCK,
-    },
-#endif
     .as =
         {
             .as_alt0 =
@@ -574,7 +548,7 @@ struct configuration_descriptor {
             .as_eac3 = USB_CARRIER_ALT(AUDIO_ALT_EAC3,
                 AUDIO_FORMAT_III_AC3 | AUDIO_FORMAT_III_DTS_I |
                 AUDIO_FORMAT_III_DTS_II | AUDIO_FORMAT_III_DTS_III, 3,
-                AUDIO_CONTROL_ID_EAC3_INPUT, AUDIO_EAC3_MAX_PACKET_SIZE),
+                AUDIO_CONTROL_ID_INPUT, AUDIO_EAC3_MAX_PACKET_SIZE),
 #endif
             #endif
         },